# 🆕 2025 FREE PROVIDERS REVOLUTION: 7 New Providers, 3x Capacity Increase!