… docs/2025_FREE_PROVIDERS_UPDATE.md
… docs/FREE_TIER_OPTIMIZATION.md
… src/modules/VisionAnalysisModule.ts
… src/renderer/components/ConfigPanel.tsx